ExternalStorageProvider: Add support for root and external storage
Google decided to only show the emulated (fuse) internal storage in
ExternalStorageProvider. This provider is used to display storages
in DocumentsUI, the Kitkat file picker, that can be used by apps.

It would be a shame if you can't pick the pictures that are on your
external SDcard :)

Change-Id: I96a8d75ccc911f6f0e7f091c20aaefd510dc3327

Source - https://gerrit.omnirom.org/#/c/2040/