fb: add general torch support for our app
till now we used an settings overlay. At least all our devices
without torch do not compile the package at all. So check
if package exists if user eg want to setup shortcuts to
torch in settings. Due that our app is needed for torch
and the user will not delete it this is an acceptable way.

see @TorchConstants and @SlimActions

Change-Id: I208fb310fbd0596dd9956d6a2c7fe106859bb4d6